refactor(core): split create-app.ts into http, command, scheduler modules
Verified: precommit-verify ✕
