+++ packages/core/src/app/create-app.ts
@@ -1,50 +1,26 @@
-import { Hono } from 'hono';
-import { match, P } from 'ts-pattern';
+import type { Hono } from 'hono';
 
 import { createContainer } from '../internal/container';
 import { LifecycleManager } from '../lifecycle';
-import { buildRoutes, warmupControllers } from '../internal/route-builder';
-import type { ErrorHandlerClass, ErrorHandlerInstance, RequestContext } from '../middleware/types';
-import { createSchedulerRunner, type SchedulerRunner } from '../scheduler/runner';
+import type { SchedulerRunner } from '../scheduler/runner';
 import type { ConfigClass } from '../config';
 import { findRootConfigToken } from '../config/token';
-import { DefaultErrorHandler } from '../http/default-error-handler';
-import { getCommandMetadata } from '../command/metadata';
 import type { CommandClass } from '../command/types';
 
+import { initializeHttp, type InitializeHttpOptions } from './http';
+import { validateCommands } from './command';
+import { registerScheduler } from './scheduler';
 import type {
   App,
   CreateAppOptions,
-  HttpOptions,
   ReadyOptions,
   ReadyResult,
   ControllerClass,
-  SchedulerClass,
 } from './types';
 
 type AnyConfigClass = ConfigClass<object>;
 type AnyConstructorClass = new (...args: never[]) => object;
 type Resolver = { get: <T extends object>(cls: new (...args: never[]) => T) => T };
-
-const createErrorHandler =
-  (errorHandlers: readonly ErrorHandlerInstance[]) =>
-  async (err: Error, c: RequestContext): Promise<Response> => {
-    for (const handler of errorHandlers) {
-      const result = await handler.onError(err, c);
-      if (result) return result;
-    }
-    throw new Error('Unreachable: DefaultErrorHandler should always return a response');
-  };
-
-const resolveErrorHandler = (cls: ErrorHandlerClass, resolver: Resolver): ErrorHandlerInstance => {
-  const instance: ErrorHandlerInstance = resolver.get(cls);
-  return instance;
-};
-
-const resolveErrorHandlers = (
-  classes: readonly ErrorHandlerClass[],
-  resolver: Resolver,
-): ErrorHandlerInstance[] => classes.map((cls) => resolveErrorHandler(cls, resolver));
 
 const instantiateConfigs = (
   configs: readonly AnyConstructorClass[] | undefined,
@@ -53,55 +29,6 @@ const instantiateConfigs = (
   for (const configClass of configs ?? []) {
     resolver.get(configClass);
   }
-};
-
-const registerScheduler = (
-  schedulers: readonly SchedulerClass[] | undefined,
-  resolver: Resolver,
-  lifecycle: LifecycleManager,
-): SchedulerRunner | undefined => {
-  if (!schedulers || schedulers.length === 0) return undefined;
-  const runner = createSchedulerRunner(schedulers, resolver);
-  lifecycle.register(runner);
-  return runner;
-};
-
-type SetupHonoOptions = {
-  readonly httpOptions: HttpOptions;
-  readonly resolver: Resolver;
-  readonly lifecycle: LifecycleManager;
-};
-
-const setupHono = (options: SetupHonoOptions): Hono => {
-  const { httpOptions, resolver, lifecycle } = options;
-  const hono = new Hono({ strict: false });
-  const userHandlers = resolveErrorHandlers(httpOptions.errorHandlers ?? [], resolver);
-  const defaultHandler = resolver.get(DefaultErrorHandler);
-  const errorHandlers = [...userHandlers, defaultHandler];
-  hono.onError(createErrorHandler(errorHandlers));
-  buildRoutes({
-    hono,
-    controllers: httpOptions.controllers,
-    resolver,
-    lifecycle,
-    globalMiddlewares: httpOptions.middlewares ?? [],
-  });
-  return hono;
-};
-
-const validateCommands = (commands: readonly CommandClass[]): Map<string, CommandClass> => {
-  const commandMap = new Map<string, CommandClass>();
-  for (const cls of commands) {
-    const meta = getCommandMetadata(cls);
-    if (!meta) {
-      throw new Error(`Command class ${cls.name} is missing @Command decorator`);
-    }
-    if (commandMap.has(meta.name)) {
-      throw new Error(`Duplicate command name: ${meta.name}`);
-    }
-    commandMap.set(meta.name, cls);
-  }
-  return commandMap;
 };
 
 type BuiltApp = {
@@ -129,73 +56,12 @@ const applyOverrides = (
 };
 
 type LifecycleResult = { ok: true } | { ok: false; cleanup: () => Promise<void> };
-type HttpResult = { hono: Hono } | { error: Error; cleanup: () => Promise<void> };
 
 const initializeLifecycle = async (lifecycle: LifecycleManager): Promise<LifecycleResult> =>
   lifecycle
     .startupPending()
     .then((): LifecycleResult => ({ ok: true }))
     .catch((): LifecycleResult => ({ ok: false, cleanup: () => lifecycle.shutdown() }));
-
-const toHttpError = (e: unknown): Error => (e instanceof Error ? e : new Error(String(e)));
-
-type InitializeHttpOptions = {
-  readonly httpOptions: HttpOptions;
-  readonly resolver: Resolver;
-  readonly lifecycle: LifecycleManager;
-};
-
-const initializeHttpSetup = (options: InitializeHttpOptions): HttpResult =>
-  match(
-    (() => {
-      try {
-        return { hono: setupHono(options) };
-      } catch (e) {
-        return { error: toHttpError(e), cleanup: () => options.lifecycle.shutdown() };
-      }
-    })(),
-  )
-    .with({ hono: P._ }, (r) => r)
-    .with({ error: P._ }, (r) => r)
-    .exhaustive();
-
-const initializeHttpWarmup = async (options: InitializeHttpOptions): Promise<HttpResult> =>
-  warmupControllers(options.httpOptions.controllers, options.resolver, options.lifecycle)
-    .then((): HttpResult => ({ hono: setupHono(options) }))
-    .catch(
-      (e): HttpResult => ({ error: toHttpError(e), cleanup: () => options.lifecycle.shutdown() }),
-    );
-
-const handleHttpSetupResult = async (result: HttpResult): Promise<Hono> =>
-  match(result)
-    .with({ hono: P._ }, (r) => r.hono)
-    .with({ error: P._ }, async (r) => {
-      await r.cleanup();
-      throw r.error;
-    })
-    .exhaustive();
-
-const handleHttpWarmupResult = async (result: HttpResult): Promise<void> => {
-  await match(result)
-    .with({ hono: P._ }, () => Promise.resolve())
-    .with({ error: P._ }, async (r) => {
-      await r.cleanup();
-      throw r.error;
-    })
-    .exhaustive();
-};
-
-const initializeHttp = async (options: InitializeHttpOptions, warmup: boolean): Promise<Hono> => {
-  const setupResult = initializeHttpSetup(options);
-  const hono = await handleHttpSetupResult(setupResult);
-
-  if (warmup) {
-    const warmupResult = await initializeHttpWarmup(options);
-    await handleHttpWarmupResult(warmupResult);
-  }
-
-  return hono;
-};
 
 const buildAppInternal = async (buildOptions: BuildAppInternalOptions): Promise<BuiltApp> => {
   const { appOptions, configOverrides, warmup, commandMap } = buildOptions;
@@ -213,7 +79,10 @@ const buildAppInternal = async (buildOptions: BuildAppInternalOptions): Promise<
   }
 
   const hono = appOptions.http
-    ? await initializeHttp({ httpOptions: appOptions.http, resolver, lifecycle }, warmup)
+    ? await initializeHttp(
+        { httpOptions: appOptions.http, resolver, lifecycle } as InitializeHttpOptions,
+        warmup,
+      )
     : undefined;
 
   return {
